Rename root package to SkyForgeBot and update config
