+++ SkyForgeBot/__init__.py
@@ -1,0 +1,2 @@
+"""SkyForgeBot package."""
+from .bot import SkyForgeBot
