+++ SkyForgeBot/bot.cfg
@@ -2,14 +2,14 @@
 # Path to loadout config. Can use relative path from here.
 looks_config = ./appearance.cfg
 
-# Path to python file. Can use relative path from here.
-python_file = ./bot.py
+# Python module path for the bot entry point.
+python_module = SkyForgeBot.bot
 requirements_file = ./requirements.txt
 
-logo_file = ./necto_logo.png
+logo_file = ./logo.png
 
 # Name of the bot in-game
-name = Necto
+name = SkyForgeBot
 
 # The maximum number of ticks per second that your bot wishes to receive.
 maximum_tick_rate_preference = 120
@@ -20,13 +20,13 @@ maximum_tick_rate_preference = 120
 developer = Rolv, Soren, and several contributors
 
 # Short description of the bot
-description = Necto is the official RLGym community bot, trained using PPO with workers run by people all around the world.
+description = SkyForgeBot is the official RLGym community bot, trained using PPO with workers run by people all around the world.
 
 # Fun fact about the bot
-fun_fact = Necto uses an attention mechanism, commonly used for text understanding, to support any number of players
+fun_fact = SkyForgeBot uses an attention mechanism, commonly used for text understanding, to support any number of players
 
 # Link to github repository
-github = https://github.com/Rolv-Arild/Necto
+github = https://github.com/Rolv-Arild/SkyForgeBot
 
 # Programming language
 language = rlgym
